backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
sqlalchemy
psycopg2-binary
